scripts/7.neural_networks/requirements.txt
opencv-python
matplotlib
pandas
numpy
h5py
pillow
tqdm
einops
torch
torchvision 
lifelines
scikit-learn
histolab
pathlib
pyyaml
webdataset
optuna
mlflow
seaborn
bidirectional-cross-attention